[WIP] Add integrated QA and telemetry for floor 6 slice 9 (#4187)
Merge-Train-PR: 4187
Merge-Train-Original-Head: 66ca21236ed7f435fc376de1afb23f710a86e247

+++ src/core/components.ts
@@ -401,7 +401,7 @@ export function createComponentStores(maxEntities = DEFAULT_MAX_ENTITIES) {
       /** Strikes landed on the outer wall by THIS ram instance. */
       strikes: new Uint16Array(maxEntities),
       /** `world.elapsedMs` of the last landed strike (0 = never). */
-      lastStrikeMs: new Float32Array(maxEntities),
+      lastStrikeMs: new Float64Array(maxEntities),
     },
     /**
      * Floor 5 · Ratings Ram escort-route waypoint marker. Carries no `Team`
